Fix: Cast dev.ProductID to int in HardwareHandler (CS0266)

+++ Services/HardwareHandler.cs
@@ -32,7 +32,7 @@ namespace GearOS.Services
                             Name = pName,
                             Vendor = dev.Manufacturer ?? "Générique",
                             VendorId = dev.VendorID,
-                            ProductId = dev.ProductID,
+                            ProductId = (int)dev.ProductID,
                             Category = n.Contains("MOUSE") ? "Souris" : n.Contains("KEYBOARD") ? "Clavier" : "Périphérique"
                         });
                     }
